melhorando o metodo batalha na classe arena

+++ src/org/unijorge/core/Ataque.java
@@ -9,7 +9,7 @@ public class Ataque {
     private int dano;
     private String nome;
     private String tipo;    
-    private double pBaseAtaque;
+    private int pBaseAtaque;
 
     //não precisa por o nome do tipo mais é feito automaticamente
     /**
@@ -19,7 +19,7 @@ public class Ataque {
      * @param tipo
      * @param pBaseAtaque
      */
-    public Ataque(int dano, String nome, String tipo, double pBaseAtaque){
+    public Ataque(int dano, String nome, String tipo, int pBaseAtaque){
         this.dano = dano;
         this.nome = nome;
         this.tipo = retornaTipoDoAtaque().get(0).toString();
@@ -32,7 +32,7 @@ public class Ataque {
      * @param nome
      * @param pBaseAtaque
      */
-    public Ataque(int dano, String nome, double pBaseAtaque){
+    public Ataque(int dano, String nome, int pBaseAtaque){
         this.dano = dano;
         this.nome = nome;
         this.pBaseAtaque = pBaseAtaque;
@@ -95,7 +95,7 @@ public class Ataque {
      * Retorna o ponto base do ataque
      * @return
      */
-    public double getpBaseAtaque() {
+    public int getpBaseAtaque() {
         return this.pBaseAtaque;
     }
 
@@ -103,7 +103,7 @@ public class Ataque {
      * ajusta o pbase ataque
      * @param pBaseAtaque
      */
-    public void setpBaseAtaque(double pBaseAtaque) {
+    public void setpBaseAtaque(int pBaseAtaque) {
         this.pBaseAtaque = pBaseAtaque;
     }
